update triangles docs and start API refactor
